factor xen-debug out of xen so we can use Xen,Debug.level in log-web and browser-loader
