+++ platform/log-node.js
@@ -5,9 +5,11 @@
 // subject to an additional IP rights grant found at
 // http://polymer.github.io/PATENTS.txt
 
+import {Debug} from '../modalities/dom/components/xen/xen-debug.js';
+
 // TODO(wkorman): Incorporate debug levels. Consider outputting
 // preamble in the specified color via ANSI escape codes. Consider
-// sharing with similar log factory logic in `xen.js`.
+// sharing with similar log factory logic in `xen.js`. See `log-web.js`.
 const logFactory = (preamble, color, log='log') => {
   return console[log].bind(console, `(${preamble})`);
 };
