+++ platform/log-web.js
@@ -5,9 +5,10 @@
 // subject to an additional IP rights grant found at
 // http://polymer.github.io/PATENTS.txt
 
-const logFactory = (preamble, color, log='log') => {
-  return console[log].bind(console, `%c${preamble}`,
-      `background: ${color}; color: white; padding: 1px 6px 2px 7px; border-radius: 6px;`);
-};
+import {Debug} from '../modalities/dom/components/xen/xen-debug.js';
 
-export {logFactory};
+const _nopFactory = () => () => {};
+const _logFactory = (preamble, color, log='log') => console[log].bind(console, `%c${preamble}`,
+      `background: ${color}; color: white; padding: 1px 6px 2px 7px; border-radius: 6px;`);
+
+export const logFactory = Debug.level < 1 ? _nopFactory : _logFactory;
